% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_autoria.R
\name{coletar_dados_autorias}
\alias{coletar_dados_autorias}
\title{Coletar Dados de Autoria de Vários Senadores}
\usage{
coletar_dados_autorias(codigos_senadores)
}
\arguments{
\item{codigos_senadores}{Vetor de códigos dos senadores para os quais os dados de autoria serão coletados.}
}
\value{
Um dataframe contendo as informações de autoria para todos os senadores fornecidos.
}
\description{
Esta função coleta dados de autoria de matérias para uma lista de senadores a partir da URL fornecida.
}
\examples{
# Exemplo de uso com códigos de senadores
codigos <- c(4981, 1234)
dados <- coletar_dados_autorias(codigos)
print(dados)
}
