% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_autorias.R
\name{coletar_dados_varios_parlamentares}
\alias{coletar_dados_varios_parlamentares}
\title{Coleta de Dados de Parlamentares no Senado}
\usage{
coletar_dados_varios_parlamentares(cod_parlamentares)
}
\arguments{
\item{cod_parlamentares}{Um vetor de strings contendo os códigos dos parlamentares para os quais os dados devem ser coletados.}
}
\value{
Um data frame contendo as informações das comissões associadas a cada parlamentar, incluindo:
\itemize{
  \item \code{CodigoParlamentar}: Código do parlamentar.
  \item \code{Sigla}: Sigla da comissão.
  \item \code{Comissao}: Nome completo da comissão.
  \item \code{Nome}: Nome do parlamentar.
  \item \code{Codigo}: Código do parlamentar.
  \item \code{SiglaPartido}: Sigla do partido ao qual o parlamentar pertence.
  \item \code{URL}: URL do perfil do parlamentar no site do Senado.
  \item \code{Foto}: URL da foto oficial do parlamentar.
  \item \code{Uf}: Unidade Federativa (UF) do parlamentar.
  \item \code{Quantidade}: Quantidade de matérias atribuídas ao parlamentar na comissão.
}
}
\description{
Esta função coleta dados de parlamentares no Senado Federal usando a API de dados abertos.
Ela permite que você colete informações sobre as comissões associadas a um ou mais parlamentares.
}
\examples{
# Coletar dados de múltiplos parlamentares
codigos_parlamentares <- c("5672", "1234", "5678")
dados_resultantes <- coletar_dados_varios_parlamentares(codigos_parlamentares)
print(dados_resultantes)

}
