% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_medidas_provisorias_tramitacao.R
\name{coletar_medidas_provisorias_em_tramitacao}
\alias{coletar_medidas_provisorias_em_tramitacao}
\title{Coletar Medidas Provisórias em Tramitação}
\usage{
coletar_medidas_provisorias_em_tramitacao()
}
\value{
Um dataframe contendo informações sobre medidas provisórias em tramitação.
O dataframe possui as seguintes colunas:
\describe{
  \item{Link}{Link para acessar mais informações sobre a medida provisória.}
  \item{Matéria}{Título da medida provisória.}
  \item{Ementa}{Resumo ou descrição da medida provisória.}
  \item{Prazo}{Prazo de vigência da medida provisória, especificando os prazos de 60 e 120 dias.}
  \item{Status}{Status da medida provisória, indicando se está em tramitação.}
}
}
\description{
Esta função coleta dados sobre medidas provisórias em tramitação do site do Congresso Nacional do Brasil.
}
\examples{
mpv_em_tramitacao <- coletar_medidas_provisorias_em_tramitacao()
}
