% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_medidas_provisorias_encerradas.R
\name{coletar_medidas_provisorias_encerradas}
\alias{coletar_medidas_provisorias_encerradas}
\title{Coletar Medidas Provisórias Encerradas}
\usage{
coletar_medidas_provisorias_encerradas(numero_ultima_pagina)
}
\arguments{
\item{numero_ultima_pagina}{Número da última página a ser coletada.}
}
\value{
Um dataframe contendo informações sobre medidas provisórias encerradas.
O dataframe possui as seguintes colunas:
\describe{
  \item{Link}{Link para acessar mais informações sobre a medida provisória.}
  \item{MPV}{Número da medida provisória.}
  \item{Título}{Título da medida provisória.}
  \item{Ementa}{Resumo ou descrição da medida provisória.}
  \item{Prazo de 60 dias}{Data limite para validade da medida provisória.}
  \item{Prazo de emendas}{Data limite para submissão de emendas.}
}
}
\description{
Esta função coleta dados sobre medidas provisórias encerradas do site do Congresso Nacional do Brasil.
}
\references{
Site do Congresso Nacional do Brasil: \url{https://www.congressonacional.leg.br}
}
\seealso{
\code{\link{coletar_medidas_provisorias_em_tramitacao}}
}
