% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_votos_senadores.R
\name{coletar_votacoes_multiplos}
\alias{coletar_votacoes_multiplos}
\title{Coleta dados de votações de múltiplos senadores em múltiplos anos}
\usage{
coletar_votacoes_multiplos(codigos_senadores, anos)
}
\arguments{
\item{codigos_senadores}{Vetor de códigos de senadores.}

\item{anos}{Vetor de anos para a coleta de votações.}
}
\value{
Um data frame com as votações de todos os senadores e anos especificados.
}
\description{
Coleta dados de votações de múltiplos senadores em múltiplos anos
}
\examples{
\dontrun{
codigos_senadores <- c("5386", "1234", "5678")
anos <- c(2023, 2024)
coletar_votacoes_multiplos(codigos_senadores, anos)
}
}
