% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convidados_reuniao.R
\name{convidados_reuniao}
\alias{convidados_reuniao}
\title{Extrair dados dos convidados de reuniões do Senado}
\usage{
convidados_reuniao(codigos_reunioes)
}
\arguments{
\item{codigos_reunioes}{Vetor contendo os códigos das reuniões}
}
\value{
Um DataFrame com os dados dos convidados de todas as reuniões empilhados
}
\description{
Esta função extrai dados dos convidados de reuniões do Senado Federal
com base nos códigos fornecidos e os empilha em um único DataFrame.
}
