% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_dados_academicos.R
\name{dados_academicos}
\alias{dados_academicos}
\title{Obter dados acadêmicos de senadores com base nos códigos parlamentares}
\usage{
dados_academicos(codigos)
}
\arguments{
\item{codigos}{Vetor de códigos parlamentares.}
}
\value{
Um dataframe com os dados acadêmicos.
}
\description{
Esta função realiza requisições para obter informações acadêmicas de senadores
com base em códigos parlamentares fornecidos.
}
\examples{
codigos <- c(5672, 5386)
dados <- dados_academicos(codigos)

}
\references{
Informações sobre os senadores são obtidas da API de dados abertos do Senado Brasileiro.
}
\seealso{
\code{\link{obter_dados_senadores_legislatura}}
}
\keyword{API}
\keyword{acadêmicos}
\keyword{dados}
\keyword{senador}
