% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_vetos.R
\name{dados_vetos}
\alias{dados_vetos}
\title{Fun\u00e7\u00e3o para extrair os dados de v\u00e1rias URLs}
\usage{
dados_vetos(urls)
}
\arguments{
\item{urls}{Um vetor de URLs das p\u00e1ginas a serem analisadas.}
}
\value{
Um dataframe contendo os dados extra\u00eddos de todas as URLs fornecidas.
Se n\u00e3o houver tabela em alguma das p\u00e1ginas, retorna uma mensagem de aviso.
}
\description{
Esta fun\u00e7\u00e3o realiza o scraping de v\u00e1rias p\u00e1ginas HTML contendo informa\u00e7\u00f5es sobre vetos do Congresso Nacional
e extrai os dados dos itens vetados.
}
\examples{
urls <- c("https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16110",
          "https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16049",
          "https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16029")
dados <- dados_vetos(urls)
}
