% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reunioes_notas.R
\name{extract_meeting_data}
\alias{extract_meeting_data}
\title{Extração de Dados de Reunião com Base no Código}
\usage{
extract_meeting_data(codigo_reuniao)
}
\arguments{
\item{codigo_reuniao}{Vetor de códigos de reunião.}
}
\value{
Um dataframe contendo informações da reunião.
}
\description{
Esta função extrai dados de uma reunião com base no código fornecido.
}
\examples{
# Exemplo de uso
codigo_reuniao <- df_reunioes$codigo
df_final <- extract_meeting_data(codigo_reuniao)
print(df_final)

}
