% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_notas_taquigraficas.R
\name{extrair_notas_taquigraficas}
\alias{extrair_notas_taquigraficas}
\title{Extrai o conte\u00fado de uma p\u00e1gina de notas taquigr\u00e1ficas.}
\usage{
extrair_notas_taquigraficas(codigos_reuniao)
}
\arguments{
\item{codigos_reuniao}{Vetor de c\u00f3digos das reuni\u00f5es para acessar as p\u00e1ginas de notas.}
}
\value{
Um dataframe contendo os c\u00f3digos de reuni\u00e3o e os respectivos conte\u00fados das p\u00e1ginas de notas, ou NULL se ocorrer um erro.
}
\description{
Esta fun\u00e7\u00e3o acessa uma p\u00e1gina de notas taquigr\u00e1ficas do Senado Federal
e extrai o conte\u00fado da mesma.
}
\examples{
codigos <- c(12071, 12072, 12073)
df_conteudos <- extrair_notas_taquigraficas(codigos)


}
