% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_pronunciamentos.R
\name{extrair_pronunciamentos_multi}
\alias{extrair_pronunciamentos_multi}
\title{Extrai dados de pronunciamentos de múltiplos parlamentares em diferentes anos}
\usage{
extrair_pronunciamentos_multi(codigos_parlamentares, anos)
}
\arguments{
\item{codigos_parlamentares}{Um vetor de códigos dos parlamentares (números inteiros).}

\item{anos}{Um vetor de anos (números inteiros) para os quais os pronunciamentos serão extraídos.}
}
\value{
Um dataframe contendo os dados de pronunciamentos de múltiplos parlamentares.
Se nenhum dado estiver disponível, retorna NULL.
}
\description{
Esta função extrai dados de pronunciamentos de múltiplos parlamentares para os anos fornecidos.
Os dados são extraídos da página do Senado Federal do Brasil.
}
\details{
A função acessa as páginas do Senado para coletar informações sobre
pronunciamentos, incluindo data, tipo, casa, partido/UF, resumo e links dos pronunciamentos.
A função percorre várias páginas para cada parlamentar e ano, com um limite máximo de 100 páginas.
}
\examples{
\dontrun{
  codigos <- c(5672, 5386)
  anos <- c(2023, 2024)
  dados_multi <- extrair_pronunciamentos_multi(codigos_parlamentares = codigos, anos = anos)
}

}
