% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_situacao_tramitacao_materia.R
\name{extrair_situacoes_tramitacao}
\alias{extrair_situacoes_tramitacao}
\title{Extrair as situa\u00e7\u00f5es da tramita\u00e7\u00e3o}
\usage{
extrair_situacoes_tramitacao(
  url = "https://legis.senado.leg.br/dadosabertos/dados/ListaSituacoes.xml"
)
}
\arguments{
\item{url}{URL do arquivo XML que cont\u00e9m as informa\u00e7\u00f5es das situa\u00e7\u00f5es da tramita\u00e7\u00e3o. Padr\u00e3o \u00e9 "https://legis.senado.leg.br/dadosabertos/dados/ListaSituacoes.xml".}
}
\value{
Um dataframe contendo as situa\u00e7\u00f5es da tramita\u00e7\u00e3o.
}
\description{
Esta fun\u00e7\u00e3o realiza o scraping de uma p\u00e1gina XML contendo informa\u00e7\u00f5es sobre as situa\u00e7\u00f5es da tramita\u00e7\u00e3o das mat\u00e9rias e extrai os dados relevantes.
}
\examples{
# Extrair as situa\u00e7\u00f5es da tramita\u00e7\u00e3o a partir da URL padr\u00e3o
df_materias_tramitacao <- extrair_situacoes_tramitacao()

}
