% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_tipo_comparecimento.R
\name{extrair_tipos_comparecimento}
\alias{extrair_tipos_comparecimento}
\title{Extrair os tipos de comparecimento em votação de uma URL específica}
\usage{
extrair_tipos_comparecimento(
  url = "https://legis.senado.leg.br/dadosabertos/dados/ListaTiposComparecimento.xml"
)
}
\arguments{
\item{url}{A URL do XML contendo os tipos de comparecimento em votação. Por padrão, é utilizada a URL "https://legis.senado.leg.br/dadosabertos/dados/ListaTiposComparecimento.xml".}
}
\value{
Um dataframe contendo os tipos de comparecimento em votação.
}
\description{
Esta função realiza o scraping de uma página XML contendo informações sobre os tipos de comparecimento em votação e extrai os dados relevantes.
}
\examples{
# Extrair os tipos de comparecimento em votação a partir da URL padrão
extrair_tipos_comparecimento()



}
