% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afastados.R
\name{get_afastados}
\alias{get_afastados}
\title{Coleta os dados dos Senadores afastados}
\usage{
get_afastados()
}
\value{
Um dataframe contendo os dados dos senadores afastados.
}
\description{
Coleta os dados dos Senadores afastados
}
\examples{
df_senadores_afastados <- get_afastados()

}
