% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_comissoes_basico.R
\name{get_dados_comissoes}
\alias{get_dados_comissoes}
\title{Coleta dados básicos das comissões permanentes}
\usage{
get_dados_comissoes()
}
\value{
Um dataframe contendo dados básicos das comissões permanentes.
}
\description{
Esta função realiza a raspagem de dados básicos das comissões permanentes do site do Senado.
}
\examples{
dados_comissoes <- get_dados_comissoes()

}
