% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reunioes.R
\name{get_reunioes}
\alias{get_reunioes}
\title{Obtém informações sobre reuniões do Senado Federal}
\usage{
get_reunioes(ano, mes)
}
\arguments{
\item{ano}{Vetor contendo os anos desejados.}

\item{mes}{Vetor contendo os meses desejados.}
}
\value{
Um dataframe contendo informações sobre as reuniões.
}
\description{
Esta função faz requisições à API de dados abertos do Senado Federal para obter informações sobre
reuniões de acordo com os anos e meses especificados.
}
\examples{
\dontrun{
# Exemplo de uso com vetores
ano <- c(2022, 2023)
mes <- c(08, 09, 10, 11, 12)
df_reunioes <- get_reunioes(ano = rep(ano, each = length(mes)), mes = rep(mes, length(ano)))
}
}
\seealso{
[API de dados abertos do Senado Federal](https://www.senado.leg.br/dadosabertos/docs/resource_agendaReuniao.html)
}
