% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_comissao_enriquecida.R
\name{info_comissao_enriquecida}
\alias{info_comissao_enriquecida}
\title{Enriquece as informações das comissões permanentes}
\usage{
info_comissao_enriquecida(urls, tipo)
}
\arguments{
\item{urls}{Vetor de URLs para extrair informações detalhadas.}

\item{tipo}{Tipo de comissão.}
}
\value{
Um dataframe contendo informações detalhadas sobre as comissões permanentes.
}
\description{
Esta função extrai informações detalhadas sobre as comissões permanentes do site do Senado
}
