% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reunioes_notas.R
\name{info_dados_reuniao}
\alias{info_dados_reuniao}
\title{Extração de Dados de Reunião}
\usage{
info_dados_reuniao(codigo_reuniao)
}
\arguments{
\item{codigo_reuniao}{Vetor de códigos de reunião.}
}
\value{
Um dataframe contendo informações da reunião.
}
\description{
Esta função extrai dados de uma reunião com base no código fornecido.
}
\examples{
codigo_reuniao <- df_reunioes$codigo
df_final <- info_dados_reuniao(codigo_reuniao)
print(df_final)

}
