% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reunioes_notas.R
\name{info_dados_reuniao_nota}
\alias{info_dados_reuniao_nota}
\title{Extração de Dados de Reunião}
\usage{
info_dados_reuniao_nota(codigo_reuniao)
}
\arguments{
\item{codigo_reuniao}{Vetor de códigos de reunião (numérico ou character).}
}
\value{
Um dataframe contendo informações das reuniões e respectivas notas taquigráficas.
}
\description{
Esta função extrai dados de uma ou mais reuniões com base nos códigos fornecidos.
}
\examples{
\dontrun{
info_dados_reuniao_nota(c(1, 2, 3))

}
}
