% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_materias_assuntos.R
\name{info_materia_assuntos}
\alias{info_materia_assuntos}
\title{Lista de Assuntos}
\usage{
info_materia_assuntos(
  url = "https://legis.senado.leg.br/dadosabertos/dados/ListaAssuntos.xml"
)
}
\arguments{
\item{url}{URL do XML contendo os dados dos assuntos. Por padrão, é usado o URL fornecido pelo Senado Federal.}
}
\value{
Um dataframe contendo os detalhes dos assuntos legislativos.
}
\description{
Esta função realiza o scraping de um XML contendo informações sobre assuntos legislativos do Senado Federal do Brasil.
Os dados extraídos incluem código do assunto, assunto geral, assunto específico, data de início e data de fim.
}
\examples{
df_assuntos = info_materia_assuntos()

}
