% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_vetos.R
\name{info_vetos}
\alias{info_vetos}
\title{Extrair detalhes de vetos a partir da URL específica contendo informações sobre vetos}
\usage{
info_vetos(pages = 1)
}
\arguments{
\item{pages}{O número de páginas a serem processadas. Padrão é 1.}
}
\value{
Um dataframe contendo os detalhes dos vetos.
}
\description{
Esta função realiza scraping de uma página contendo informações sobre vetos e extrai detalhes como número do veto, link associado, ementa, data de sobresta, matéria vetada e norma gerada.
}
\examples{
# Exemplo de uso
dados_vetos <- info_vetos(pages = 3)

}
