% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_medidas_provisorias.R
\name{medidas_provisorias}
\alias{medidas_provisorias}
\title{Coletar Medidas Provisórias}
\usage{
medidas_provisorias(inicio_pagina, fim_pagina)
}
\arguments{
\item{inicio_pagina}{Número da primeira página a ser coletada}

\item{fim_pagina}{Número da última página a ser coletada}
}
\value{
Um dataframe com dados sobre medidas provisórias em andamento e encerradas
}
\description{
Esta função coleta dados sobre medidas provisórias em andamento e encerradas
do site do Congresso Nacional do Brasil.
}
\examples{
mpv_data <- medidas_provisorias(1, 3)
}
