% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_notas_taquigraficas.R
\name{notas_taquigraficas}
\alias{notas_taquigraficas}
\title{Extrai o conteúdo de uma página de notas taquigráficas.}
\usage{
notas_taquigraficas(codigos_reuniao)
}
\arguments{
\item{codigos_reuniao}{Vetor de códigos das reuniões para acessar as páginas de notas.}
}
\value{
Um dataframe contendo os códigos de reunião e os respectivos conteúdos das páginas de notas, ou NULL se ocorrer um erro.
}
\description{
Esta função acessa uma página de notas taquigráficas do Senado Federal
e extrai o conteúdo da mesma.
}
\examples{
codigos <- c(12071, 12072, 12073)
df_conteudos <- notas_taquigraficas(codigos)
}
