% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_partidos.R
\name{obter_dados_partidos}
\alias{obter_dados_partidos}
\title{Obtém dados dos partidos dos senadores a partir da API do Senado}
\usage{
obter_dados_partidos()
}
\value{
Um dataframe contendo dados dos partidos.
}
\description{
Esta função realiza uma requisição à API do Senado para obter informações sobre os partidos dos senadores.
}
\examples{
df_partidos <- obter_dados_partidos()

}
