% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_senadores_legislatura.R
\name{obter_dados_senadores_legislatura}
\alias{obter_dados_senadores_legislatura}
\title{Obter dados dos senadores por legislatura}
\usage{
obter_dados_senadores_legislatura(legislatura_inicio, legislatura_fim)
}
\arguments{
\item{legislatura_inicio}{N\u00famero da legislatura de in\u00edcio.}

\item{legislatura_fim}{N\u00famero da legislatura de fim.}
}
\value{
Um dataframe com os dados dos senadores.
}
\description{
Esta fun\u00e7\u00e3o realiza requisi\u00e7\u00f5es HTTP para obter informa\u00e7\u00f5es dos senadores
com base nas legislaturas fornecidas.
}
\examples{
df_senadores_legislatura <- obter_dados_senadores_legislatura(47, 56)

}
\references{
Informa\u00e7\u00f5es sobre os senadores s\u00e3o obtidas da API de dados abertos do Senado Brasileiro.
}
\keyword{API}
\keyword{legislatura}
\keyword{senador}
