% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorias_parlamentar.R
\name{obter_relatorias_parlamentares}
\alias{obter_relatorias_parlamentares}
\title{Obter relatorias de parlamentares}
\usage{
obter_relatorias_parlamentares(codigos_parlamentares, anos)
}
\arguments{
\item{codigos_parlamentares}{Vetor de códigos de parlamentares.}

\item{anos}{Vetor de anos.}
}
\value{
Um dataframe consolidado com todas as relatorias.
}
\description{
Esta função obtém as relatorias de parlamentares com base nos códigos e anos fornecidos.
}
\examples{
# Exemplo de uso
df_relatorias <- obter_relatorias_parlamentares(codigos_parlamentares = c(6331, 1234), anos = c(2022, 2023))
print(df_relatorias)

}
