% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_apartes.R
\name{processar_xml_apartes}
\alias{processar_xml_apartes}
\title{Processa XML de Apartes de Senadores}
\usage{
processar_xml_apartes(codigos)
}
\arguments{
\item{codigos}{Um vetor de códigos de parlamentares para os quais os dados de
apartes serão extraídos.}
}
\value{
Um dataframe contendo os dados dos apartes, com as seguintes colunas:
\describe{
  \item{CodigoParlamentar}{Código do parlamentar}
  \item{NomeParlamentar}{Nome do parlamentar}
  \item{CodigoPronunciamento}{Código do pronunciamento}
  \item{TipoUsoPalavra}{Tipo de uso da palavra}
  \item{DataPronunciamento}{Data do pronunciamento}
  \item{SiglaPartidoParlamentarNaData}{Sigla do partido do parlamentar na data}
  \item{UfParlamentarNaData}{UF do parlamentar na data}
  \item{SiglaCasaPronunciamento}{Sigla da casa do pronunciamento}
  \item{NomeCasaPronunciamento}{Nome da casa do pronunciamento}
  \item{TextoResumo}{Resumo do texto do pronunciamento}
  \item{Indexacao}{Indexação do pronunciamento}
  \item{UrlTexto}{URL do texto do pronunciamento}
  \item{UrlTextoBinario}{URL do texto binário do pronunciamento}
  \item{CodigoSessao}{Código da sessão plenária}
  \item{NomeCasaSessao}{Nome da casa da sessão}
  \item{DataSessao}{Data da sessão}
  \item{HoraInicioSessao}{Hora de início da sessão}
  \item{OradorCodigo}{Código do orador}
  \item{OradorNome}{Nome do orador}
  \item{PublicacaoData}{Data da publicação}
  \item{PublicacaoUrl}{URL da publicação}
}
}
\description{
Esta função acessa a URL da API do Senado para cada código de parlamentar fornecido,
extrai e processa os dados dos apartes associados a cada parlamentar.
Retorna um dataframe com informações sobre os apartes, incluindo dados do
parlamentar e detalhes dos apartes.
}
\examples{
\dontrun{
# Exemplo de uso com códigos fictícios
codigos <- c("4763", "1234")
df_apartes <- processar_xml_apartes(codigos)
}

}
