% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_liderancas.R
\name{processar_xml_liderancas}
\alias{processar_xml_liderancas}
\title{Processa XML de Lideran\u00e7as de Senadores}
\usage{
processar_xml_liderancas(codigos)
}
\arguments{
\item{codigos}{Um vetor de c\u00f3digos de parlamentares para os quais os dados de
lideran\u00e7as ser\u00e3o extra\u00eddos.}
}
\value{
Um dataframe contendo os dados das lideran\u00e7as, com as seguintes colunas:
\describe{
  \item{CodigoParlamentar}{C\u00f3digo do parlamentar}
  \item{NomeParlamentar}{Nome do parlamentar}
  \item{UnidadeLideranca}{Unidade da lideran\u00e7a}
  \item{CodigoPartido}{C\u00f3digo do partido}
  \item{SiglaPartido}{Sigla do partido}
  \item{NomePartido}{Nome do partido}
  \item{SiglaCasaLideranca}{Sigla da casa de lideran\u00e7a}
  \item{NomeCasaLideranca}{Nome da casa de lideran\u00e7a}
  \item{DescricaoTipoLideranca}{Descri\u00e7\u00e3o do tipo de lideran\u00e7a}
  \item{DataDesignacao}{Data de designa\u00e7\u00e3o}
  \item{DataFim}{Data de t\u00e9rmino}
  \item{CodigoBloco}{C\u00f3digo do bloco}
  \item{SiglaBloco}{Sigla do bloco}
  \item{NomeBloco}{Nome do bloco}
  \item{ApelidoBloco}{Apelido do bloco}
  \item{NumeroOrdemViceLideranca}{N\u00famero de ordem da vice-lideran\u00e7a}
}
}
\description{
Esta fun\u00e7\u00e3o acessa a URL da API do Senado para cada c\u00f3digo de parlamentar fornecido,
extrai e processa os dados de lideran\u00e7as associados a cada parlamentar.
Retorna um dataframe com informa\u00e7\u00f5es sobre as lideran\u00e7as, incluindo dados do
parlamentar e detalhes das lideran\u00e7as.
}
\examples{
\dontrun{
# Exemplo de uso com c\u00f3digos fict\u00edcios
codigos <- c("123456", "654321")
df_liderancas <- processar_xml_liderancas(codigos)
}

}
