% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_mandatos.R
\name{processar_xml_mandatos}
\alias{processar_xml_mandatos}
\title{Processar XMLs de Mandatos de Senadores}
\usage{
processar_xml_mandatos(codigos)
}
\arguments{
\item{codigos}{Um vetor de códigos de senadores para os quais os dados de mandatos serão obtidos.}
}
\value{
Um dataframe contendo os dados dos mandatos dos senadores especificados. Cada linha do dataframe
        representa um mandato e inclui informações como o código do mandato, a unidade federativa, e as datas
        de início e fim das legislaturas.
}
\description{
Esta função processa os XMLs disponíveis na URL do Senado para obter dados dos mandatos de múltiplos senadores.
Para cada código de senador fornecido, a função faz o download e a análise do XML correspondente,
extraindo informações sobre os mandatos do senador e combinando-os em um único dataframe.
}
\examples{
\dontrun{
df_mandatos_senadores <- processar_xml_mandatos(c("5322", "1234"))
}
}
