% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_profissoes.R
\name{processar_xml_profissoes}
\alias{processar_xml_profissoes}
\title{Processa XML de Profissões de Senadores}
\usage{
processar_xml_profissoes(codigos)
}
\arguments{
\item{codigos}{Um vetor de códigos de parlamentares para os quais os dados de
profissões serão extraídos.}
}
\value{
Um dataframe contendo os dados das profissões, com as seguintes colunas:
\describe{
  \item{CodigoParlamentar}{Código do parlamentar}
  \item{NomeParlamentar}{Nome do parlamentar}
  \item{NomeProfissao}{Nome da profissão}
  \item{IndicadorAtividadePrincipal}{Indicador se a profissão é a atividade principal}
}
}
\description{
Esta função acessa a URL da API do Senado para cada código de parlamentar fornecido,
extrai e processa os dados de profissões associados a cada parlamentar.
Retorna um dataframe com informações sobre as profissões, incluindo dados do
parlamentar e detalhes das profissões.
}
\examples{
\dontrun{
# Exemplo de uso com códigos fictícios
codigos <- c("4981", "1234")
df_profissoes <- processar_xml_profissoes(codigos)
}

}
