% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorias_parlamentar.R
\name{relatorias_parlamentares}
\alias{relatorias_parlamentares}
\title{Obter relatorias de parlamentares}
\usage{
relatorias_parlamentares(codigos_parlamentares, anos)
}
\arguments{
\item{codigos_parlamentares}{Vetor de códigos de parlamentares.}

\item{anos}{Vetor de anos.}
}
\value{
Um dataframe consolidado com todas as relatorias.
}
\description{
Esta função obtém as relatorias de parlamentares com base nos códigos e anos fornecidos.
}
\examples{
# Exemplo de uso
df_relatorias <- relatorias_parlamentares(codigos_parlamentares = c(5386, 5012), anos = c(2022, 2023))
print(df_relatorias)

}
