% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_orientacao_votacao.R
\name{coletar_orientacao_votacao}
\alias{coletar_orientacao_votacao}
\title{Coletar dados de votação do Senado Federal}
\usage{
coletar_orientacao_votacao(data_sessao)
}
\arguments{
\item{data_sessao}{Um vetor contendo as datas das sessões no formato "AAAA-MM-DD".}
}
\value{
Um data frame contendo os dados de votação, incluindo código de votação, data e hora, partido e voto.
}
\description{
Esta função coleta os dados de votação do Senado Federal a partir de uma lista de datas de sessão.
}
\examples{
\dontrun{
# Exemplo de uso com vetor de datas de sessão
dados_votacao <- coletar_orientacao_votacao(df_vot$data_sessao)
}
}
