% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_votos_senadores.R
\name{coletar_votacoes_multiplos}
\alias{coletar_votacoes_multiplos}
\title{Coleta dados de vota\u00e7\u00f5es de m\u00faltiplos senadores em m\u00faltiplos anos}
\usage{
coletar_votacoes_multiplos(codigos_senadores, anos)
}
\arguments{
\item{codigos_senadores}{Vetor de c\u00f3digos de senadores.}

\item{anos}{Vetor de anos para a coleta de vota\u00e7\u00f5es.}
}
\value{
Um data frame com as vota\u00e7\u00f5es de todos os senadores e anos especificados.
}
\description{
Coleta dados de vota\u00e7\u00f5es de m\u00faltiplos senadores em m\u00faltiplos anos
}
\examples{
codigos_senadores <- c("5386", "1234", "5678")
anos <- c(2023, 2024)
coletar_votacoes_multiplos(codigos_senadores, anos)
}
