% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_classificacao_materia.R
\name{extrair_classificacoes_materia}
\alias{extrair_classificacoes_materia}
\title{Extrair as classificações de matéria de uma URL específica}
\usage{
extrair_classificacoes_materia(
  url = "https://legis.senado.leg.br/dadosabertos/dados/ListaClassificacoesMateria.xml"
)
}
\arguments{
\item{url}{A URL do XML contendo as classificações de matéria. Por padrão, é utilizada a URL "https://legis.senado.leg.br/dadosabertos/dados/ListaClassificacoesMateria.xml".}
}
\value{
Um dataframe contendo as classificações de matéria.
}
\description{
Esta função realiza o scraping de uma página XML contendo informações sobre as classificações de matéria e extrai os dados relevantes.
}
\examples{
df_classificacao_materias <- extrair_classificacoes_materia()
}
