% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_composicao_comissao_web.R
\name{extrair_dados_comissao_senador}
\alias{extrair_dados_comissao_senador}
\title{Extrai dados das comissões de um senador}
\usage{
extrair_dados_comissao_senador(codigo_senador)
}
\arguments{
\item{codigo_senador}{O código do senador.}
}
\value{
Um data frame contendo os dados das comissões do senador.
}
\description{
Esta função extrai os dados das comissões de um senador com base no seu código parlamentar.
}
\examples{
\dontrun{
codigo_senador <- "5008"
extrair_dados_comissao_senador(codigo_senador)
}

}
