% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_vetos.R
\name{extrair_detalhes_vetos}
\alias{extrair_detalhes_vetos}
\title{Extrair detalhes de vetos a partir da URL específica contendo informações sobre vetos}
\usage{
extrair_detalhes_vetos(pages = 1)
}
\arguments{
\item{pages}{O número de páginas a serem processadas.}
}
\value{
Um dataframe contendo os detalhes dos vetos.
}
\description{
Esta função realiza scraping de uma página contendo informações sobre vetos e extrai detalhes como número do veto, link associado, ementa, data de sobresta, matéria vetada e norma gerada.
}
