% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_detalhes_vetos.R, R/info_vetos.R
\name{extrair_detalhes_vetos}
\alias{extrair_detalhes_vetos}
\title{Função para extrair os dados de várias URLs}
\usage{
extrair_detalhes_vetos(pages = 1)

extrair_detalhes_vetos(pages = 1)
}
\arguments{
\item{pages}{O número de páginas a serem processadas.}

\item{urls}{Um vetor de URLs das páginas a serem analisadas.}
}
\value{
Um dataframe contendo os dados extraídos de todas as URLs fornecidas.
Se não houver tabela em alguma das páginas, retorna uma mensagem de aviso.

Um dataframe contendo os detalhes dos vetos.
}
\description{
Esta função realiza o scraping de várias páginas HTML contendo informações sobre vetos do Congresso Nacional
e extrai os dados da tabela de partes vetadas.

Esta função realiza scraping de uma página contendo informações sobre vetos e extrai detalhes como número do veto, link associado, ementa, data de sobresta, matéria vetada e norma gerada.
}
\examples{
urls <- c("https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16110",
          "https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16049",
          "https://www.congressonacional.leg.br/materias/vetos/-/veto/detalhe/16029")
dados <- extrair_detalhes_vetos(urls)

dados_vetos <- extrair_detalhes_vetos(pages = 3)
head(dados_vetos)

}
