% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_materias_tramitando.R
\name{extrair_materias_tramitando}
\alias{extrair_materias_tramitando}
\title{Extrair informações das matérias em tramitação de uma URL específica}
\usage{
extrair_materias_tramitando(
  url = "https://legis.senado.leg.br/dadosabertos/materia/tramitando"
)
}
\arguments{
\item{url}{A URL do XML contendo as matérias em tramitação. Por padrão, é utilizada a URL "https://legis.senado.leg.br/dadosabertos/materia/tramitando".}
}
\value{
Um dataframe contendo informações das matérias em tramitação.
}
\description{
Esta função realiza o scraping de uma página XML contendo informações sobre as matérias em tramitação e extrai os dados relevantes.
}
\examples{
# Extrair informações das matérias em tramitação a partir da URL padrão
extrair_materias_tramitando()

}
