% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_pronunciamentos.R
\name{extrair_pronunciamentos_multi}
\alias{extrair_pronunciamentos_multi}
\title{Extrai dados de pronunciamentos de múltiplos parlamentares em diferentes anos}
\usage{
extrair_pronunciamentos_multi(codigos_parlamentares, anos)
}
\arguments{
\item{codigos_parlamentares}{Vetor de códigos dos parlamentares.}

\item{anos}{Vetor de anos para os quais os pronunciamentos serão extraídos.}
}
\value{
Um dataframe contendo os dados de pronunciamentos de múltiplos parlamentares.
Se nenhum dado estiver disponível, retorna NULL.
}
\description{
Esta função extrai dados de pronunciamentos de múltiplos parlamentares para os anos fornecidos.
}
\examples{
codigo <- c(5672, 5386)
ano <- c(2023, 2024)
dados_multi <- extrair_pronunciamentos_multi(codigos_parlamentares = codigo, anos = ano)

}
