% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_situacao_tramitacao_materia.R
\name{extrair_situacoes_tramitacao}
\alias{extrair_situacoes_tramitacao}
\title{Extrair as situações da tramitação}
\usage{
extrair_situacoes_tramitacao(
  url = "https://legis.senado.leg.br/dadosabertos/dados/ListaSituacoes.xml"
)
}
\value{
Um dataframe contendo as situações da tramitação.
}
\description{
Esta função realiza o scraping de uma página XML contendo informações sobre as situações da tramitação das matérias e extrai os dados relevantes.
}
\examples{
# Extrair as situações da tramitação a partir da URL padrão
df_materias_tramitação = extrair_situacoes_tramitacao()

}
