% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_tipo_decisao_materia.R
\name{extrair_tipos_decisao}
\alias{extrair_tipos_decisao}
\title{Extrair os tipos de decisão de uma URL específica}
\usage{
extrair_tipos_decisao(
  url = "https://legis.senado.leg.br/dadosabertos/dados/ListaTiposDecisao.xml"
)
}
\value{
Um dataframe contendo os tipos de decisão.
}
\description{
Esta função realiza o scraping sobre os tipos de decisão e extrai os dados relevantes.
}
\examples{
# Extrair os tipos de decisão
df_tipos_decisao <- extrair_tipos_decisao()
}
