% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_votos_nominais.R
\name{extrair_votacoes_nominais_por_ano}
\alias{extrair_votacoes_nominais_por_ano}
\title{Extrai as informações das votações nominais por ano}
\usage{
extrair_votacoes_nominais_por_ano(anos)
}
\arguments{
\item{anos}{Vetor contendo os anos das votações a serem extraídas}
}
\value{
Um data frame com as informações das votações nominais
}
\description{
Esta função extrai as informações das votações nominais e retorna um data frame.
}
\examples{
Executa a função para extrair as votações nominais
anos = c(2022, 2023)
dados_votacoes <- extrair_votacoes_nominais_por_ano(anos)

}
