% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_agenda.R
\name{info_agenda}
\alias{info_agenda}
\title{Raspar os dados da agenda para várias datas}
\usage{
info_agenda(anos, meses, dias)
}
\arguments{
\item{anos}{Vetor contendo os anos desejados.}

\item{meses}{Vetor contendo os meses desejados.}

\item{dias}{Vetor contendo os dias desejados.}
}
\value{
Um dataframe contendo os dados da agenda para as datas especificadas.
}
\description{
Esta função raspa os dados da agenda de uma página da web para várias datas e
combina os resultados em um único dataframe.
}
\examples{
resultado <- info_agenda(anos = c(2023, 2024), meses = c(1, 2), dias = 10:15)
}
