% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_agenda.R
\name{info_agenda}
\alias{info_agenda}
\title{Extrair dados da agenda do Congresso Nacianal, do Senado federal e Câmara dos Deputados}
\usage{
info_agenda(anos, meses, dias)
}
\arguments{
\item{anos}{Vetor de anos.}

\item{meses}{Vetor de meses.}

\item{dias}{Vetor de dias.}
}
\value{
Um dataframe contendo os dados da agenda do Congresso Nacional.
}
\description{
Esta função extrai dados da agenda do Congresso Nacional para os anos, meses e dias especificados.
}
\examples{
resultado <- info_agenda(anos = c(2023, 2024), meses = c(1, 2), dias = 10:15)

}
