% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_atualizacao_materias.R
\name{info_atualizacao_materias}
\alias{info_atualizacao_materias}
\title{Lista de Matérias Atualizadas}
\usage{
info_atualizacao_materias(
  url = "https://legis.senado.leg.br/dadosabertos/materia/atualizadas"
)
}
\arguments{
\item{url}{URL do XML contendo os dados das matérias atualizadas. Por padrão, é usado o URL fornecido pelo Senado Federal.}
}
\value{
Um dataframe contendo os detalhes das matérias legislativas atualizadas recentemente.
}
\description{
Esta função realiza o scraping de um XML contendo informações sobre matérias legislativas atualizadas recentemente no Senado Federal do Brasil.
Os dados extraídos incluem código da matéria, sigla e nome da casa legislativa, tipo e descrição do subtipo da matéria, número e ano da matéria, descrição do objetivo do processo,
descrição da matéria, indicador de tramitação, identificação do processo, ementa da matéria, indexação da matéria, indicador complementar, datas de apresentação e leitura da matéria,
classificações da matéria (se disponível), e informações sobre atualizações recentes da matéria.
}
\examples{
df_materias_atualizacao <- info_atualizacao_materias()

}
