% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_atualizacao_materias.R
\name{info_atualizacao_materias}
\alias{info_atualizacao_materias}
\title{Lista de Matérias Atualizadas}
\usage{
info_atualizacao_materias(
  url = "https://legis.senado.leg.br/dadosabertos/materia/atualizadas"
)
}
\value{
Um dataframe contendo os detalhes das matérias legislativas atualizadas recentemente.
}
\description{
Esta função realiza o scraping de informações sobre matérias legislativas atualizadas recentemente no Senado Federal do Brasil.
}
\examples{
df_materias_atualizacao <- info_atualizacao_materias()

}
