% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reunioes.R
\name{info_dados_reuniao_comissao}
\alias{info_dados_reuniao_comissao}
\title{Extrair dados das reuniões das comissões parlamentares}
\usage{
info_dados_reuniao_comissao(codcol, data_inicio, data_fim)
}
\arguments{
\item{codcol}{Vetor de códigos das comissões.}

\item{data_inicio}{Data de início do período desejado no formato "AAAA-MM-DD".}

\item{data_fim}{Data de fim do período desejado no formato "AAAA-MM-DD".}
}
\value{
Um dataframe contendo as informações extraídas das notas taquigráficas das comissões.
}
\description{
Esta função extrai dados das reuniões de uma comissão específica do Senado Federal.
}
\examples{
codcol <- c(54, 38, 2614, 34)
data_inicio <- "2023-01-01"
data_fim <- "2023-12-01"
df_reunioes <- info_dados_reuniao_comissao(codcol, data_inicio, data_fim)

}
