% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reunioes_notas.R
\name{info_dados_reuniao_nota}
\alias{info_dados_reuniao_nota}
\title{Extra\u00e7\u00e3o de Dados de Reuni\u00e3o}
\usage{
info_dados_reuniao_nota(codigo_reuniao)
}
\arguments{
\item{codigo_reuniao}{Vetor de c\u00f3digos de reuni\u00e3o (num\u00e9rico ou character).}
}
\value{
Um dataframe contendo informa\u00e7\u00f5es das reuni\u00f5es e respectivas notas taquigr\u00e1ficas.
}
\description{
Esta fun\u00e7\u00e3o extrai dados de uma ou mais reuni\u00f5es com base nos c\u00f3digos fornecidos.
}
\examples{
info_dados_reuniao_nota(c(1, 2, 3))

}
