% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reunioes_notas.R
\name{info_dados_reuniao_nota}
\alias{info_dados_reuniao_nota}
\title{Extração de Dados de Reunião}
\usage{
info_dados_reuniao_nota(codigo_reuniao)
}
\arguments{
\item{codigo_reuniao}{Vetor de códigos de reunião.}
}
\value{
Um dataframe contendo informações da reunião.
}
\description{
Esta função extrai dados de uma reunião com base no código fornecido.
}
\examples{
df_reunioes <- data.frame(codigo = 1:10, outra_coluna = letters[1:10])
info_dados_reuniao_nota(df_reunioes)

}
