% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_materias_assuntos.R
\name{info_materia_temas}
\alias{info_materia_temas}
\title{Lista de Assuntos}
\usage{
info_materia_temas(
  url = "https://legis.senado.leg.br/dadosabertos/dados/ListaAssuntos.xml"
)
}
\value{
Um dataframe contendo os detalhes dos assuntos legislativos.
O dataframe possui as seguintes colunas:
\describe{
  \item{Codigo}{Código do assunto.}
  \item{AssuntoGeral}{Assunto geral ao qual o assunto pertence.}
  \item{AssuntoEspecifico}{Assunto específico.}
  \item{DataInicio}{Data de início da validade do assunto.}
  \item{DataFim}{Data de término da validade do assunto, se aplicável.}
}
}
\description{
Esta função realiza o scraping de um XML contendo informações sobre assuntos legislativos do Senado Federal do Brasil.
}
\examples{
df_assuntos <- info_materia_temas()

}
