% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_materia_legislatura_atual.R
\name{materias_legislatura_atual}
\alias{materias_legislatura_atual}
\title{Lista de Matérias da Legislatura Atual}
\usage{
materias_legislatura_atual()
}
\value{
Um dataframe contendo os detalhes das matérias legislativas.
O dataframe possui as seguintes colunas:
\describe{
  \item{CodigoMateria}{Código da matéria.}
  \item{SiglaCasaIdentificacaoMateria}{Sigla da casa legislativa identificadora da matéria.}
  \item{NomeCasaIdentificacaoMateria}{Nome da casa legislativa identificadora da matéria.}
  \item{SiglaSubtipoMateria}{Sigla do subtipo da matéria.}
  \item{NumeroMateria}{Número da matéria.}
  \item{AnoMateria}{Ano da matéria.}
  \item{DescricaoIdentificacaoMateria}{Descrição da identificação da matéria.}
  \item{IndicadorTramitando}{Indicador de tramitação da matéria.}
  \item{IdentificacaoProcesso}{Identificação do processo da matéria.}
}
}
\description{
Esta função realiza o scraping de um XML contendo informações sobre matérias legislativas da legislatura atual do Senado Federal do Brasil.
Os dados extraídos incluem o código da matéria, sigla e nome da casa legislativa, subtipo da matéria, número e ano, descrição, indicação de tramitação e identificação do processo.
}
\examples{
df_materias <- materias_legislatura_atual()

}
