% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_materia_legislatura_atual.R
\name{materias_legislatura_atual}
\alias{materias_legislatura_atual}
\title{Lista de Matérias da Legislatura Atual}
\usage{
materias_legislatura_atual()
}
\value{
Um dataframe contendo os detalhes das matérias legislativas.
}
\description{
Esta função realiza o scraping de um XML contendo informações sobre matérias legislativas da legislatura atual do Senado Federal do Brasil.
Os dados extraídos incluem o código da matéria, sigla e nome da casa legislativa, subtipo da matéria, número e ano, descrição, indicação de tramitação e identificação do processo.
}
\examples{

df_materias <- materias_legislatura_atual()

}
