% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_medidas_provisorias.R
\name{medidas_provisorias}
\alias{medidas_provisorias}
\title{Coletar Medidas Provisórias}
\usage{
medidas_provisorias(inicio_pagina, fim_pagina)
}
\arguments{
\item{inicio_pagina}{Number of the first page to scrape}

\item{fim_pagina}{Number of the last page to scrape}
}
\value{
A dataframe with data from provisional measures in progress and closed
}
\description{
This function scrapes data from provisional measures in progress and closed
from the Brazilian National Congress website.
}
\examples{
mpv_data <- medidas_provisorias(1, 3)
}
