% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_composicao_comissoes_api.R
\name{obter_dados_comissoes_parlamentares}
\alias{obter_dados_comissoes_parlamentares}
\title{Obter dados das comissões de parlamentares}
\usage{
obter_dados_comissoes_parlamentares(codigos_parlamentares)
}
\arguments{
\item{codigos_parlamentares}{Vetor contendo os códigos dos parlamentares.}
}
\value{
Um dataframe contendo os dados das comissões que parlamentares são ou foram membros.
}
\description{
Esta função obtém dados das comissões de parlamentares com base em seus códigos.
}
\examples{
codigos_senadores <- df_parlamentares$CodigoParlamentar
df_comissoes_parlamentares <- obter_dados_comissoes_parlamentares(codigos_senadores)

}
