% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_partidos.R
\name{obter_dados_partidos}
\alias{obter_dados_partidos}
\title{Obt\u00e9m dados dos partidos dos senadores a partir da API do Senado}
\usage{
obter_dados_partidos()
}
\value{
Um dataframe contendo dados dos partidos.
}
\description{
Esta fun\u00e7\u00e3o realiza uma requisi\u00e7\u00e3o \u00e0 API do Senado para obter informa\u00e7\u00f5es sobre os partidos.
}
\examples{
df_partidos <- obter_dados_partidos()

}
