% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_liderancas.R
\name{processar_xml_liderancas}
\alias{processar_xml_liderancas}
\title{Processa XML de Lideranças de Senadores}
\usage{
processar_xml_liderancas(codigos)
}
\arguments{
\item{codigos}{Um vetor de códigos de parlamentares para os quais os dados de
lideranças serão extraídos.}
}
\value{
Um dataframe contendo os dados das lideranças, com as seguintes colunas:
\describe{
  \item{CodigoParlamentar}{Código do parlamentar}
  \item{NomeParlamentar}{Nome do parlamentar}
  \item{UnidadeLideranca}{Unidade da liderança}
  \item{CodigoPartido}{Código do partido}
  \item{SiglaPartido}{Sigla do partido}
  \item{NomePartido}{Nome do partido}
  \item{SiglaCasaLideranca}{Sigla da casa de liderança}
  \item{NomeCasaLideranca}{Nome da casa de liderança}
  \item{DescricaoTipoLideranca}{Descrição do tipo de liderança}
  \item{DataDesignacao}{Data de designação}
  \item{DataFim}{Data de término}
  \item{CodigoBloco}{Código do bloco}
  \item{SiglaBloco}{Sigla do bloco}
  \item{NomeBloco}{Nome do bloco}
  \item{ApelidoBloco}{Apelido do bloco}
  \item{NumeroOrdemViceLideranca}{Número de ordem da vice-liderança}
}
}
\description{
Esta função acessa a URL da API do Senado para cada código de parlamentar fornecido,
extrai e processa os dados de lideranças associados a cada parlamentar.
Retorna um dataframe com informações sobre as lideranças, incluindo dados do
parlamentar e detalhes das lideranças.
}
\examples{
\dontrun{
# Exemplo de uso com códigos fictícios
codigos <- c("123456", "654321")
df_liderancas <- processar_xml_liderancas(codigos)
}

}
