% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_profissoes.R
\name{processar_xml_profissoes}
\alias{processar_xml_profissoes}
\title{Processa XML de Profiss\u00f5es de Senadores}
\usage{
processar_xml_profissoes(codigos)
}
\arguments{
\item{codigos}{Um vetor de c\u00f3digos de parlamentares para os quais os dados de
profiss\u00f5es ser\u00e3o extra\u00eddos.}
}
\value{
Um dataframe contendo os dados das profiss\u00f5es, com as seguintes colunas:
\describe{
  \item{CodigoParlamentar}{C\u00f3digo do parlamentar}
  \item{NomeParlamentar}{Nome do parlamentar}
  \item{NomeProfissao}{Nome da profiss\u00e3o}
  \item{IndicadorAtividadePrincipal}{Indicador se a profiss\u00e3o \u00e9 a atividade principal}
}
}
\description{
Esta fun\u00e7\u00e3o acessa a URL da API do Senado para cada c\u00f3digo de parlamentar fornecido,
extrai e processa os dados de profiss\u00f5es associados a cada parlamentar.
Retorna um dataframe com informa\u00e7\u00f5es sobre as profiss\u00f5es, incluindo dados do
parlamentar e detalhes das profiss\u00f5es.
}
\examples{
\dontrun{
# Exemplo de uso com c\u00f3digos fict\u00edcios
codigos <- c("4981", "1234")
df_profissoes <- processar_xml_profissoes(codigos)
}

}
